requirements.txt
numpy
Pillow
pathlib
click